Use data_get to retrieve value from resource
Using Laravel's `data_get` helper allows nested values to be used for the dependencies, which is important for packages like Nova Page Manager where all the fields are stored in a `data` JSON column (that is casted into an array in the model). This allows dependencies on field attributes such as `data->type`. This PR does not introduce any breaking changes and does not affect existing setups.